requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
aniso8601==10.0.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
argcomplete==3.5.3
asgiref==3.8.1
attrs==24.3.0
azure-core==1.32.0
azure-monitor-opentelemetry-exporter==1.0.0b33
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
easyocr==1.7.2
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
fixedint==0.1.6
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google_search_results==2.4.1
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-redis==0.1.2
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.2.10
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
mdurl==0.1.2
mem0ai==0.1.44
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.6
onnxruntime==1.20.1
openai==1.59.7
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
paddleocr==2.9.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
posthog==3.8.3
promptflow==1.17.1
promptflow-core==1.17.1
promptflow-devkit==1.17.1
promptflow-tools==1.4.0
promptflow-tracing==1.17.1
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydash==7.0.7
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-doctr==0.10.0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
python-ulid==2.7.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.2
RapidFuzz==3.11.0
redis==5.2.1
redisvl==0.3.8
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tesseract==0.1.3
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
waitress==3.0.2
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
